server/requirements.txt
flask>=2.3.3
flask-cors>=4.0.0
flask-restx>=1.3.0
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
stripe>=7.0.0
PyJWT>=2.8.0
cryptography>=41.0.0

